+++ dump/config/atmosphere/retrieval_amv_viirs.yaml
