+++ dump/config/atmosphere/retrieval_osw_ascat.yaml
