+++ dump/scripts/atmosphere/radiance_avhrr.py
@@ -5,7 +5,7 @@ import numpy as np
 import bufr
 from bufr.obs_builder import ObsBuilder, add_main_functions, map_path
 
-MAPPING_PATH = map_path('avhrr.yaml')
+MAPPING_PATH = map_path('radiance_avhrr.yaml')
 
 
 class BufrGsrasrObsBuilder(ObsBuilder):
