+++ dump/scripts/atmosphere/radiance_cris-fsr.py
@@ -7,7 +7,7 @@ import bufr
 from bufr.obs_builder import ObsBuilder, add_main_functions, map_path
 
 
-MAPPING_PATH = map_path('cris-fsr.yaml')
+MAPPING_PATH = map_path('radiance_cris-fsr.yaml')
 
 
 class BufrCrisObsBuilder(ObsBuilder):
