+++ dump/scripts/atmosphere/radiance_iasi.py
@@ -7,7 +7,7 @@ import bufr
 from bufr.obs_builder import ObsBuilder, add_main_functions, map_path
 
 
-MAPPING_PATH = map_path('iasi.yaml')
+MAPPING_PATH = map_path('radiance_iasi.yaml')
 
 
 class BufrIasiObsBuilder(ObsBuilder):
