+++ dump/scripts/atmosphere/retrieval_osw_ascat.py
@@ -8,7 +8,7 @@ from bufr.obs_builder import ObsBuilder, add_main_functions, map_path, add_dummy
 from bufr.transforms import compute_wind_components
 
 
-MAPPING_PATH = map_path('scatwnd_ascat.yaml')
+MAPPING_PATH = map_path('retrieval_osw_ascat.yaml')
 
 
 class BufrAscatObsBuilder(ObsBuilder):
